Lookup fuel types, fix session, cyclic imports
