+++ api/app/utils/cffdrs.py
@@ -27,7 +27,7 @@ def rate_of_spread(fuel_type: str, isi: float, bui: float):
     if fuel_type is None or isi is None or bui is None:
         return None
     # pylint: disable=protected-access, no-member
-    result = cffdrs._ROScalc(FUELTYPE=fuel_type, ISI=isi, BUI=bui, CBH=6, SFC=1.5, PC=2)
+    result = cffdrs._ROScalc(FUELTYPE="C7", ISI=isi, BUI=bui, CBH=6, SFC=1.5, PC=2)
     return result[0]
 
 # Args:
